G33TaskManager/TaskManager/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==0.7
distlib==0.3.9
Django==4.2.7
django-config==0.1.4
django-crispy-forms==2.0
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kaggle==1.6.17
MarkupSafe==3.0.2
oauthlib==3.2.2
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-slugify==8.0.4
python-socketio==5.12.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
stripe==11.6.0
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
